# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "纸喵软件系列之服务端",
        "title": "纸喵 wechat API",
        "termsOfService": "http://zhimiao.org",
        "contact": {
            "name": "API Support",
            "url": "http://tools.zhimiao.org",
            "email": "mail@xiaoliu.org"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/miniprogram/Config": {
            "post": {
                "description": "小程序配置",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "小程序控制"
                ],
                "summary": "小程序配置",
                "parameters": [
                    {
                        "description": "body参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/req.MiniAppConfigParam"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\": 1,\"msg\": \"操作成功\",\"data\": null}",
                        "schema": {
                            "$ref": "#/definitions/resp.ApiResult"
                        }
                    }
                }
            }
        },
        "/miniprogram/GetWXACodeUnlimit": {
            "get": {
                "tags": [
                    "小程序信息获取"
                ],
                "summary": "获取线上小程序码",
                "parameters": [
                    {
                        "type": "string",
                        "description": "小程序APPID",
                        "name": "MiniProgramID",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "页面",
                        "name": "Page",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "场景，不得超过32位，不得含有特殊符号",
                        "name": "Scene",
                        "in": "query"
                    }
                ]
            }
        },
        "/miniprogram/Lists": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "小程序信息获取"
                ],
                "summary": "小程序列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "排序列",
                        "name": "orderColumn",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "小程序appid",
                        "name": "appID",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "开放平台ID",
                        "name": "platformID",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "-1-授权失效 1授权成功，2审核中，3审核通过，4审核失败，5已发布",
                        "name": "state",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "排序方式，desc倒序， asc正序",
                        "name": "orderClase",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "pageSize",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "lastId",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "版本号",
                        "name": "version",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/resp.MiniAppListsVO"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "req.MiniAppConfigParam": {
            "type": "object",
            "required": [
                "appID"
            ],
            "properties": {
                "appID": {
                    "description": "小程序appid",
                    "type": "string"
                },
                "autoAudit": {
                    "description": "自动提审(升级) -1否 1是",
                    "type": "integer"
                },
                "autoRelease": {
                    "description": "自动发布-1否 1是",
                    "type": "integer"
                },
                "extConfig": {
                    "description": "小程序扩展配置，发布时会注入至ext.json",
                    "type": "string"
                },
                "mchID": {
                    "description": "支付商户号id",
                    "type": "string"
                },
                "secret": {
                    "description": "小程序secret",
                    "type": "string"
                }
            }
        },
...
        "resp.ApiResult": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "type": "object"
                },
                "msg": {
                    "type": "string"
                }
            }
        },
        "resp.MiniAppListsVO": {
            "type": "object",
            "properties": {
                "appID": {
                    "description": "小程序appid",
                    "type": "string"
                },
                "auditID": {
                    "description": "审核编号",
                    "type": "integer"
                },
                "autoAudit": {
                    "description": "自动提审(升级) -1否 1是",
                    "type": "integer"
                },
                "autoRelease": {
                    "description": "自动发布-1否 1是",
                    "type": "integer"
                },
                "createTime": {
                    "type": "string"
                },
                "extConfig": {
                    "description": "小程序扩展配置，发布时会注入至ext.json",
                    "type": "string"
                },
                "mchID": {
                    "description": "支付商户号id",
                    "type": "string"
                },
                "originalID": {
                    "description": "原始ID",
                    "type": "string"
                },
                "platformID": {
                    "description": "开放平台ID",
                    "type": "string"
                },
                "refreshToken": {
                    "description": "接口调用凭据刷新令牌",
                    "type": "string"
                },
                "secret": {
                    "description": "小程序secret",
                    "type": "string"
                },
                "state": {
                    "description": "-1-授权失效 1授权成功，2审核中，3审核通过，4审核失败，5已发布",
                    "type": "integer"
                },
                "templateListen": {
                    "description": "模板监听开发小程序(appid)",
                    "type": "string"
                },
                "updateTime": {
                    "type": "string"
                },
                "version": {
                    "description": "当前版本",
                    "type": "string"
                }
            }
        }
    }
}